// GestionTienda/src/datamatrix-module/components/orders/interfaces/orderdefault.ts
export interface IOrderDefault{
      id: string;  
      omsId: string;  
      omsUrl: string;  
      token: string;  
      connectionId: string;  
      contactPerson: string;  
      createMethodType: string;  
      factoryAddress: string;  
      factoryCountry: string;  
      factoryId: number;  
      factoryName: string;  
      productionLineId: number;  
      releaseMethodType: string;  
}